JoomlaForge Tracker #24047
Fixed bug that required a hyphen in the copyright date even though the copyright was only a single year. Current behavior will now accept the hyphen as the first part of the copyright name if present. It will also accept a second year without a hyphen.

A better fix for this is made problematic by the uncertain nature of the copyright holder name. If a copyright holder name begins with 4 numbers, it will be virtually indistinguishable from a year, rendering it difficult to decide if a hyphen should be between it.